+++ sdk/cpp/tests/test_c_api.cpp
@@ -1,28 +1,26 @@
-//////////////////////////////////////////////////////////////////
-// @file golang.cpp
-//
-// YANG Development Kit
-// Copyright 2017 Cisco Systems. All rights reserved
-//
-////////////////////////////////////////////////////////////////
-// Licensed to the Apache Software Foundation (ASF) under one
-// or more contributor license agreements.  See the NOTICE file
-// distributed with this work for additional information
-// regarding copyright ownership.  The ASF licenses this file
-// to you under the Apache License, Version 2.0 (the
-// "License"); you may not use this file except in compliance
-// with the License.  You may obtain a copy of the License at
-//
-//   http://www.apache.org/licenses/LICENSE-2.0
-//
-//  Unless required by applicable law or agreed to in writing,
-// software distributed under the License is distributed on an
-// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-// KIND, either express or implied.  See the License for the
-// specific language governing permissions and limitations
-// under the License.
-//
-//////////////////////////////////////////////////////////////////
+/*  ----------------------------------------------------------------
+ @file ydk.cpp
+
+ YDK - YANG Development Kit
+ Copyright 2017-2019 Cisco Systems. All rights reserved.
+
+ Licensed under the Apache License, Version 2.0 (the "License");
+ you may not use this file except in compliance with the License.
+ You may obtain a copy of the License at
+
+ http://www.apache.org/licenses/LICENSE-2.0
+
+ Unless required by applicable law or agreed to in writing, software
+ distributed under the License is distributed on an "AS IS" BASIS,
+ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ See the License for the specific language governing permissions and
+ limitations under the License.
+ -------------------------------------------------------------------
+ This file has been modified by Yan Gorelik, YDK Solutions.
+ All modifications in original under CiscoDevNet domain
+ introduced since October 2019 are copyrighted.
+ All rights reserved under Apache License, Version 2.0.
+ ------------------------------------------------------------------*/
 
 #include <iostream>
 
@@ -146,5 +144,29 @@ TEST_CASE( "c_api_rpc" )
     NetconfServiceProviderFree(provider);
     RepositoryFree(repo);
     CodecFree(c);
+    YDKStateFree(state);
+}
+
+TEST_CASE( "c_api_session"  )
+{
+    // Connect to Netconf server
+    YDKStatePtr state = YDKStateCreate();
+    Repository repo = RepositoryInit();
+    REQUIRE(repo!=NULL);
+    auto session = NetconfSessionInit(state, repo, "localhost", "admin", "admin", 12022,
+                                      "ssh", true, false, -1, "", "");
+    REQUIRE(session!=NULL);
+
+    // Get capabilities
+    int len = 0;
+    auto capabilities = NetconfSessionGetCapabilities(state, session, &len);
+    CHECK(len > 0);
+    cout << "Total capabilities - "<< len << endl;
+    cout << capabilities[0] << endl;
+    CapabilitiesArrayFree(capabilities, len);
+
+    // Disconnect from Netconf server
+    NetconfSessionFree(session);
+    RepositoryFree(repo);
     YDKStateFree(state);
 }